Implement iframe-based redirection
This introduces a new file gathering all iframe-related functions. These
are then simply called in the appropriate contexts across the codebase.
The introduced functions aime at:
- Redirecting the browser to an IRI in an iframe
- Letting the parent window know to what the redirect IRI the Solid Identity
Provider has  redirected the iframe to.
- The main window cleans up the iframe after it handled the redirection.
- The session triggers the iframe-based flow based on a signal. In
future commit, this signal may be emitted when an access token expires.
